requirements.txt
llama-index
python-dotenv
streamlit
streamlit-pills
llama-index-embeddings-google
llama-index-llms-gemini
chromadb
llama-index-vector-stores-chroma